Use clean-install command only if package.lock is available

+++ sw6/dev/actions/setup.sh
@@ -4,9 +4,17 @@
 ACTION: dev:composer-install
 
 if [ -f __SW6_ADMINISTRATION_ASSET_FOLDER__/package.json ]; then
+   if [ -f __SW6_ADMINISTRATION_ASSET_FOLDER__/package.lock ]; then
    npm clean-install --prefix __SW6_ADMINISTRATION_ASSET_FOLDER__;
+   else
+   npm install --prefix __SW6_ADMINISTRATION_ASSET_FOLDER__;
+   fi
    fi
 
 if [ -f __SW6_STOREFRONT_ASSET_FOLDER__/package.json ]; then
+   if [ -f __SW6_STOREFRONT_ASSET_FOLDER__/package.lock ]; then
    npm clean-install --prefix __SW6_STOREFRONT_ASSET_FOLDER__;
+   else
+   npm install --prefix __SW6_STOREFRONT_ASSET_FOLDER__;
+   fi
    fi
